+++ test/integration/climate.test.d.ts
@@ -1,3 +1,0 @@
-#!/usr/bin/env node
-export {};
-//# sourceMappingURL=climate.test.d.ts.map